+++ core/main.ts
@@ -33,7 +33,6 @@ export * from "./src/ecs/systems/ui/RenderSystem"
 export * from "./src/ecs/entities/buttons/ConnectButton"
 export * from "./src/ecs/entities/buttons/FullscreenButton"
 export * from "./src/ecs/entities/buttons/HomeButton"
-export * from "./src/ecs/entities/buttons/ShopButton"
 export * from "./src/ecs/entities/characters/Chicko"
 export * from "./src/ecs/entities/characters/Piggo"
 export * from "./src/ecs/entities/characters/Skelly"
