+++ core/src/ecs/entities/buttons/FullscreenButton.ts
@@ -4,7 +4,7 @@ export const FullscreenButton = (id: string = "fullscreenButton") => Entity({
   id: id,
   components: {
     position: Position({ x: -40, y: 5, screenFixed: true }),
-    clickable: Clickable({ active: true, width: 32, height: 30 }),
+    clickable: Clickable({ width: 32, height: 30 }),
     actions: Actions({
       click: ({ world }) => {
         if (!document.fullscreenElement) {
